Added Validators.java to encapsulate validation methods in the initial record lookup to ensure integrity as records are prepared for submission to the database.
TO DO:
- Create a qualifier, in which if the BOOLEAN Admitted is FALSE, the optional A Number should become MANDATORY, including a denial reason, and an override of the auto calculated ExitDate for certain visa classes.
I might consider elevating the Admitted Boolean higher to trigger this in an easier fashion with less headway.  As one will require certain reasons, but auto NULLing for certain classes can help with record entries (Citizens cannot be denied entry, citizens should NOT have Alien Numbers if entering with a US Passport (validating with the CountryCode) to AUTO NULL certain classes.