admissionClasses are now validated against the AdmissionClass.txt file.
The Admissions Process will immediately begin upon the flight details are entered.

Citizens with their US Passport will be expedited immediately.

Next will be Permanent Residents, who would use their ARCs for re-entry.  (if else case 1)(Foreign passport+ARC) to add in admission, class and other items and blocks denial and reasoning, immediately nullifies the exit date and mandates an Alien Number to be entered (then checked by the method (TODO: fix the method to be discretionary in certain cases of admission and MANDATORY for other cases and DENIED admissions (consider automatic A Number entry for Denials if one is not added in the instant case or the Passport number+Country combo does not return an A Number)

Next will be all other aliens applying for admission:
1. IVs applying, once admitted, will have their class passed through the method for a NULL exit date and override to nullify any removal dates. and forces their A-Number to be entered.
2. Aliens with D/S status will have the maximum stay calculated (or for the sake of simulation a certain amount of years will be forced.)
3. All other aliens with fixed maximum dates, the method will calculate and return the proposed MAX date, and the officer can override it for a date UNDER that (ideal for tourists (B2/B1) and transiting aliens on C1)
4. Aliens denied entry will have the admissionClassExitDate() bypassed.  and the officer will enter their ExitDate as their date of removal manually.